First round of file deletions and reorganization
