--- Frontend/src/Pages/Dashboard.js
@@ -1,17 +1,0 @@
-// Added in npm install react-router-dom to get the packages to change to different pages.
-import React from 'react';
-//import { useEffect, useState } from "react";
-//import { Navigate } from "react-router-dom";
-
-function Dashboard()
-{
-    return (
-        <div>
-            <h1> Welcome to Dashboard </h1>
-        </div>
-    );
-    
-
-}
-
-export default Dashboard;
