--- Frontend/src/components/Home/Home.css
@@ -1,23 +1,0 @@
-.top {
-
-    width: 100%;
-    height: 250px;
-    margin: auto;
-    margin-top: 0px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-}
-
-.top .toptext{ 
-    margin: auto;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    font-size: 5rem;
-
-}
-
-
-
