--- Frontend/src/components/Home/Home.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import './Home.css'
-const Home = () => {
-    return (
-        <div className = "home"> 
-            <div className = "top">
-                <div className = "toptext"> 
-                    
-                </div> 
-                
-            </div>
-        </div>
-       
-    )
-}
-
-export default Home
