+++ Frontend/src/components/Student/QuizContent.js
