--- Frontend/src/components/TeacherRedirect/TeacherAssessment.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import '../Mentor/PageLayout.css'
-
-function TeacherAssessment()
-{
-    return (
-        <div class="body">
-                <h1> Welcome to Create/Edit Assessment </h1> 
-        </div>
-    );
-    
-
-}
-
-export default TeacherAssessment;
