--- Frontend/src/components/TeacherRedirect/TeacherQuiz.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import '../Mentor/PageLayout.css';
-
-function TeacherQuiz()
-{
-    return (
-        <div class="body">
-            <h1> Welcome to Create/Edit Quiz </h1> 
-        </div>
-    );
-    
-
-}
-
-export default TeacherQuiz;
