--- Frontend/src/components/TeacherRedirect/TeacherTeam.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import '../Mentor/PageLayout.css';
-
-function TeacherTeam()
-{
-    return (
-        <div class="body">
-            <h1> Welcome to Team </h1> 
-        </div>
-    );
-    
-
-}
-
-export default TeacherTeam;
