--- Frontend/src/components/TeamDropdown/TeamDropdown.css
@@ -1,63 +1,0 @@
-.teamdropdown {
-    width: 100%;
-    height: 50px;
-    background-color: #648FFF;
-}
-
-.teamdropdown .teamcontainer{ 
-    max-width: 1140px;
-    height: 100%;
-    margin: auto; 
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-}
-
-.teamdropdown-menu{
-   
-    text-align: center;
-    color:black;
-    background: white;
-    position: relative;
-    cursor:pointer;
-}
-
-.teamdropdown-menu:hover{
-    color:black;
-    background: white;
-    
-}
-
-.teamdropdown-menu{
-    width: 7rem;
-    text-align: center;
-    color:black;
-    background: #648FFF;
-    position: relative;
-    cursor:pointer;
-}
-
-.teamdropdown-menu:hover{
-    color:black;
-    background: white;
-    
-}
-
-.teamdropdown-list{
-    border: 1px solid transparent;
-    position:absolute;
-    left:0;
-    width:100%;
-    background: lightgrey;
-}
-
-.teamdropdown-list>li{
-    border: 1px solid transparent;
-}
-
-.teamdropdown-list>li:hover{
-    border: 1px solid white;
-}
-
-
