--- Frontend/src/components/TeamStats/TeamStats.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const TeamStats = () => {
-    return (
-        <div>
-
-            Team Stats check
-        </div>
-    )
-}
-
-export default TeamStats
