--- Frontend/src/components/ViewScores/ViewScores.css
@@ -1,14 +1,0 @@
-
-  table {
-    border: 5px solid rgb(0, 0, 0);
-    width: 1000px;
-    height: 400px;
-  }
-    
-  th {
-    border-bottom: 1px solid rgb(98, 27, 27);
-  }
-    
-  td {
-    text-align: center;
-  }
