+++ Frontend/src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom/client';
 // eslint-disable-next-line
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import './index.css';
-import App from './components/App/App.js';
+import App from './App/App.js';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
